+++ src/app/static/src/app/components/modelOutput/ModelOutput.vue
@@ -100,6 +100,20 @@
                     :formatFunction="formatBarchartValue"
                     :showRangesInTooltips="true"
                     @update="updateComparisonPlotSelectionsAndXAxisOrder"></bar-chart-with-filters>
+                <div class="row mt-2">
+                    <div class="col-md-3"></div>
+                    <area-indicators-table class="col-md-9"
+                                        :table-data="comparisonPlotData"
+                                        :area-filter-id="areaFilterId"
+                                        :filters="comparisonPlotFilters"
+                                        :countryAreaFilterOption="countryAreaFilterOption"
+                                        :indicators="filteredComparisonPlotIndicators"
+                                        :selections="comparisonPlotSelections"
+                                        :translate-filter-labels="false"
+                                        :selectedFilterOptions="comparisonPlotSelections.selectedFilterOptions"
+                    ></area-indicators-table>
+                </div>
+                <error-alert v-if="!!comparisonPlotError" :error="comparisonPlotError"></error-alert>
             </div>
         </div>
     </div>
@@ -112,6 +126,7 @@
     import AreaIndicatorsTable from "../plots/table/AreaIndicatorsTable.vue";
     import {BarchartIndicator, Filter, FilterConfig, FilterOption} from "@reside-ic/vue-charts/src/bar/types";
     import {BarChartWithFilters} from "@reside-ic/vue-charts";
+    import ErrorAlert from "../ErrorAlert.vue";
 
     import {
         mapGettersByNames,
@@ -190,8 +205,10 @@
         filteredChoroplethIndicators: ChoroplethIndicatorMetadata[],
         filteredBarchartIndicators: BarchartIndicator[],
         filteredBubblePlotIndicators: ChoroplethIndicatorMetadata[],
+        filteredComparisonPlotIndicators: BarchartIndicator[],
         barchartFlattenedXAxisFilterOptionIds: string[]
         comparisonPlotFlattenedXAxisFilterOptionIds: string[]
+        comparisonPlotError: Error | null
     }
 
     export default Vue.extend<Data, Methods, Computed, unknown>({
@@ -236,6 +253,9 @@
                     featureLevels: state => state.shape!.filters.level_labels || []
                 }
             ),
+            ...mapStateProps<ModelCalibrateState, keyof Computed>("modelCalibrate", {
+                comparisonPlotError: state => state.comparisonPlotError
+            }),
             filteredChoroplethIndicators() {
                 return this.choroplethIndicators.filter((val: ChoroplethIndicatorMetadata) => val.indicator === this.choroplethSelections.indicatorId)
             },
@@ -247,6 +267,9 @@
                     ...this.bubblePlotIndicators.filter((val: ChoroplethIndicatorMetadata) => val.indicator === this.bubblePlotSelections.colorIndicatorId),
                     ...this.bubblePlotIndicators.filter((val: ChoroplethIndicatorMetadata) => val.indicator === this.bubblePlotSelections.sizeIndicatorId)
                 ]
+            },
+            filteredComparisonPlotIndicators() {
+                return this.comparisonPlotIndicators.filter((val: BarchartIndicator) => val.indicator === this.comparisonPlotSelections.indicatorId)
             },
             selectedTab: mapStateProp<ModelOutputState, string>("modelOutput", state => state.selectedTab),
             chartdata: mapStateProp<ModelCalibrateState, any>("modelCalibrate", state => {
@@ -296,7 +319,8 @@
             BarChartWithFilters,
             BubblePlot,
             Choropleth,
-            AreaIndicatorsTable
+            AreaIndicatorsTable,
+            ErrorAlert
         }
     })
 </script>