+++ src/app/static/src/app/components/plots/table/AreaIndicatorsTable.vue
@@ -40,6 +40,7 @@
         filters: Filter[],
         countryAreaFilterOption: NestedFilterOption,
         areaFilterId: string
+        translateFilterLabels: boolean
     }
 
     interface DisplayRow {
@@ -80,6 +81,10 @@
         },
         selections: {
             type: Object
+        },
+        translateFilterLabels: {
+            type: Boolean,
+            default: true
         }
     }
 
@@ -172,7 +177,7 @@
                 this.filtersToDisplay.forEach(value => {
                     const field: Dict<any> = {};
                     field.key = value.id
-                    field.label = i18next.t(value.label.toLowerCase(), {lng: this.currentLanguage})
+                    field.label = this.translateFilterLabels ? i18next.t(value.label.toLowerCase(), {lng: this.currentLanguage}) : value.label
                     fields.push(field)
                 })
                 this.indicators.forEach(value => {