+++ NEWS.md
@@ -1,3 +1,15 @@
+# hint 2.11.1
+
+* fix broken request password mailer
+
+# hint 2.11.0
+
+* Table for comparison barchart
+
+# hint 2.10.0
+
+* Display comparison plot error message
+
 # hint 2.9.1
 
 * Bug: HINT login is broken on dev and staging